feat: Encrypting local data with key derived from password

+++ apps/web/src/hooks/useInvites.ts
@@ -5,21 +5,24 @@ import { generateBase58Id } from "@chat/crypto";
 import { InviteMessage } from "@chat/sockets";
 import { InviteType } from "@chat/core";
 import { refreshRooms } from "@/lib/utils";
+import { useAuth } from "./useAuth";
 
 export function useInvites() {
-  const db = useDB();
+  const { db, putEncr, getAllDecr } = useDB();
   const [invites, setInvites] = useState<InviteType[]>([]);
+  const { key } = useAuth();
 
   useEffect(() => {
-    if (!db) return;
+    if (!db || !key) return;
 
     let cleanup: (() => void) | undefined;
 
     const setup = async () => {
       const client = await getSignalingClient();
-      if (!client) return;
+      console.log("client is available");
 
       const handleRoomInvite = async (msg: InviteMessage) => {
+        console.log(msg);
         const inviteId = generateBase58Id();
 
         const newInvite: InviteType = {
@@ -28,7 +31,8 @@ export function useInvites() {
           from: msg.from,
         };
 
-        await db.put("invites", newInvite);
+        const encr = await putEncr("invites", newInvite, key);
+        console.log(encr);
         setInvites((prev) => [...prev, newInvite]);
 
         console.log(`Received room invite from ${msg.from}: ${msg.room.name}`);
@@ -36,7 +40,7 @@ export function useInvites() {
 
       client.on("invite", handleRoomInvite);
 
-      const storedInvites = await db.getAll("invites");
+      const storedInvites = await getAllDecr("invites", key) as InviteType[];
       if (storedInvites) {
         const normalized = storedInvites.map((i: InviteType) => ({
           inviteId: i.inviteId ?? generateBase58Id(),
@@ -54,18 +58,18 @@ export function useInvites() {
     setup();
 
     return () => cleanup?.();
-  }, [db]);
+  }, [db, key]);
 
   const acceptInvite = async (invite: InviteType) => {
-    if (!db) return;
+    if (!db || !key) return;
 
-    await db.put("rooms", {
+    await putEncr("rooms", {
       ...invite.room,
       roomId: generateBase58Id(),
-    });
+    }, key);
 
     for(const cred of invite.room.keys) {
-      await db.put("credentials", cred);
+      await putEncr("credentials", cred, key);
     }
 
     await db.delete("invites", invite.inviteId);
